Update README.md -> zooming in on Lapply
I have worked further on the lapply funciton. I think what we had was still wrong as it still referred to Silvia's working directory whilst the TA explicitly said to remove that (as they dont have access to our local computer where we have the datafiles stored) and she suggested using the Lapply function along with their respective urls. I've followed Hannes' function when it comes to Lapply and I also had to change how to access individual datasets as the names e.g. "title_basics" was not yet created. Instead in the file it was name as [[1]]. But after this step it should be called the "title_basics" etc. And this way we dont have to use our local working directory