Scatterplot manually added...
I manually added the R codes to add the scatterplot. The code is also on my R script which is connected to Github, but when i saved the R script and chose my remote branch i cant see where i stored it along with my text so i just manually did it. But the scatterplot does seems to show that short series do get higher ratings as a first impression